+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## v2.0.0-alpha.next
+
+## v2.0.0-alpha.2
   * Update `seq` package iterators to only cancel on hard-stop.
   * Stop child processes synchronously if added to a stopped parent.
-  * `Context.WaitCtx()` returns the delegate's `Err()` joined with any errors accumulated by the stopper.
+  * `Context.WaitCtx()` returns the argument's `Err()` joined with any errors accumulated by the stopper.
 
 ## v2.0.0-alpha.1
   * Complete API overhaul. See the [migration guide](doc/migration.md)